Separate color_block source
